pyproject.toml
[project]
name = "ai-train-2026"
version = "0.1.0"
description = "AI training soft for AI course WSIZ Bielsko-Biała 2026"
readme = "README.md"
requires-python = "<3.13,>=3.12"
dependencies = [
    "loguru>=0.7.3",
    "pydantic>=2.13.4",
    # deps for torch working on RTX 5090
    "transformers<6.0.0,>=5.3.0",
    "torch==2.5.1+cu124",
    "torchvision==0.20.1+cu124",
    "torchaudio==2.5.1+cu124",
    "python-dotenv>=1.2.2",
]


[[tool.uv.index]]
url = "https://download.pytorch.org/whl/cu124"


[tool.uv]
package = false

[build-system]
requires = ["uv-build"]
build-backend = "uv_build"
